URL Restructuring and Folder Move
Moved all files out of the /laravel directory to be in the document root.

Moved files out of /public into document root to make urls not read /blogtest/public/. This is not best practice but, due to limitations with my shared hosting environment I could not place core Laravel files outside the document root. Also htaccess methods to change the URL did not work.